+++ package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "main": "src/index.js",
   "scripts": {
-    "test": "node src/js/joker.js",
+    "test": "node src/js/joker.test.js",
     "format": "prettier -w src/js/joker.js"
   },
   "repository": {
